requirements.txt
Flask==3.0.3
Jinja2==3.1.4
SQLAlchemy==2.0.30
Flask-WTF==1.2.1
gunicorn==22.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1  # 新規追加